% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl-metric.R
\name{check_metric}
\alias{check_metric}
\title{Perform sanity checks on a metric object}
\usage{
check_metric(metric)
}
\arguments{
\item{metric}{A metric table, as found in the data field of a metric object}
}
\description{
This function previously worked on metric data tables, but it now works on metric objects
(which contain all the metadata, documentation, and everything needed to plot).
}
