% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete-periods.R
\name{complete_periods}
\alias{complete_periods}
\title{Given a cumulative metric, add dates for the end of each period}
\usage{
complete_periods(metric, periods = c("month"))
}
\arguments{
\item{metric}{A metric table in wide format, containing "date" and "period" columns as
well as one or more dimensions and metric values.}

\item{periods}{Vector of periods to add: one or more of "week", "month", "quarter" or "year".}
}
\description{
Some metrics like ARR are measured cumulatively, so in order to create a bar plot
per month or quarter we need to pick the last value from each period. For example,
the ARR for January 2019 would be measured as of 2019-01-31. Analogously
to the tidyr function \code{complete()}, this adds rows representing each period
present in the data.
}
\examples{

library(dplyr)

flights <- nycflights13::flights \%>\%
  mutate(date = as.Date(ISOdate(year, month, day)))

# Include number and cumulative number of flights
cumulative_summary <- flights \%>\%
  cross_by_periods(periods = "day") \%>\%
  summarize(nb_flights = n()) \%>\%
  arrange(date) \%>\%
  mutate(cumulative_flights = cumsum(nb_flights)) \%>\%
  ungroup()

# Have periods for week and month as well, representing the end of that period
library(ggplot2)

cumulative_day_week_month <- cumulative_summary \%>\%
  complete_periods(periods = c("week", "month"))

cumulative_day_week_month \%>\%
  ggplot(aes(date, cumulative_flights, color = period)) +
  geom_point()
}
