% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl-metric.R
\name{condense_metric}
\alias{condense_metric}
\title{Condense a metric_tbl object to remove cases with multiple non-All dimensions}
\usage{
condense_metric(metric, max_dimensions = 2)
}
\arguments{
\item{metric}{A \code{tbl_metric} object}

\item{max_dimensions}{The number of (non-All) dimensions that each row
can have}
}
\description{
This reduces the size of a metrics table, by limiting the number of dimensions
that can be anything besides All at the same time. If there is a \code{min_dimensions}
field in the metric metadata, it never condenses beyond that (this is useful for some
that need multiple dimensions to be interpretable)
}
