% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions.R
\name{discard_dimensions}
\alias{discard_dimensions}
\title{Discard dimensions}
\usage{
discard_dimensions(tbl, ..., quietly = FALSE)
}
\arguments{
\item{tbl}{A metric tbl in wide format, with one or more dimensions.}

\item{...}{Dimensions to discard.}

\item{quietly}{If FALSE (default), display a message about what columns are
being discarded.}
}
\description{
This function discards specified dimensions summary by filtering only for the
attribute value "All". If no dimensions are specified, it discards all of them.
If the table is grouped, this ungroups it.
}
\examples{
library(dplyr)

mtcars_by_cyl_gear <- mtcars \%>\%
  cross_by_dimensions(cyl, gear) \%>\%
  summarize(avg_mpg = mean(mpg))

# Discard all dimensions
mtcars_by_cyl_gear \%>\%
  discard_dimensions()

# Remove dimension cyl
mtcars_by_cyl_gear \%>\%
  discard_dimensions(cyl)

# Remove all dimensions except `cyl`
mtcars_by_cyl_gear \%>\%
  discard_dimensions(-cyl)

mtcars_by_cyl_gear \%>\%
  discard_dimensions(-cyl, -gear)

mtcars_by_cyl_gear \%>\%
  discard_dimensions(-one_of("cyl", "gear"))
}
\seealso{
\code{\link[=keep_dimensions]{keep_dimensions()}}
}
