% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gather_metrics}
\alias{gather_metrics}
\title{Gather all the non-dimension columns in a wide metric table}
\usage{
gather_metrics(tbl, ..., quietly = FALSE)
}
\arguments{
\item{tbl}{A table to gather metrics from}

\item{...}{A selection of columns to gather. If empty, all numeric columns
are selected as a default}

\item{quietly}{A boolean indicating if diagnostic messages are to be printed.}
}
\description{
Gather all the non-dimension columns in a wide metric table
}
