% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-date-periods.R
\name{generate_date_periods}
\alias{generate_date_periods}
\title{Generate a table with pairings of dates and periods}
\usage{
generate_date_periods(
  start,
  end,
  periods = c("day", "week", "month", "quarter", "year"),
  windows = c(7, 28, 56),
  intervals = FALSE,
  today_date = lubridate::today()
)
}
\arguments{
\item{start}{Start date}

\item{end}{End date}

\item{periods}{A vector of calendar periods. This supports "day", "week", "month", "quarter",
and "year".}

\item{windows}{A vector of windows, each representing a # of days}

\item{intervals}{Whether a preselected set of intervals starting from today, such as
"Last Week", "Last 2 Weeks", or "All Time" should be included.}

\item{today_date}{Date to count as "today": by default, \code{lubridate::today()}.}
}
\description{
Generate a table with pairings of dates and periods
}
