% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions.R
\name{keep_dimensions}
\alias{keep_dimensions}
\title{Keep dimensions}
\usage{
keep_dimensions(tbl, ..., keep_attribute_all = FALSE, quietly = FALSE)
}
\arguments{
\item{tbl}{A metric tbl in wide format, with one or more dimensions.}

\item{...}{Dimensions to keep, as bare names
or select helpers like \code{contains()}.}

\item{keep_attribute_all}{Whether to remove the "All" level from the dimensions}

\item{quietly}{If FALSE (default), display a message about what columns are
being discarded.}
}
\description{
This function keeps specified dimensions from a wide metric tbl and discards the rest.
}
\examples{
library(dplyr)

mtcars_by_cyl_gear <- mtcars \%>\%
  cross_by_dimensions(cyl, gear) \%>\%
  summarize(avg_mpg = mean(mpg))

mtcars_by_cyl_gear \%>\%
  keep_dimensions()

mtcars_by_cyl_gear \%>\%
  keep_dimensions(cyl)

mtcars_by_cyl_gear \%>\%
  keep_dimensions(-cyl)

mtcars_by_cyl_gear \%>\%
  keep_dimensions(cyl, keep_attribute_all = TRUE)

}
\seealso{
\code{\link{discard_dimensions}}
}
