% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl-metric-group.R
\name{metric-group-s3}
\alias{metric-group-s3}
\alias{as_tibble.tbl_metric_group}
\alias{filter.tbl_metric_group}
\alias{select.tbl_metric_group}
\alias{arrange.tbl_metric_group}
\alias{mutate.tbl_metric_group}
\alias{group_by.tbl_metric_group}
\alias{summarise.tbl_metric_group}
\alias{inner_join.tbl_metric_group}
\alias{left_join.tbl_metric_group}
\alias{right_join.tbl_metric_group}
\alias{semi_join.tbl_metric_group}
\alias{anti_join.tbl_metric_group}
\alias{rename.tbl_metric_group}
\alias{count.tbl_metric_group}
\alias{transmute.tbl_metric_group}
\title{Metric dplyr S3 methods}
\usage{
\method{as_tibble}{tbl_metric_group}(x, ...)

\method{filter}{tbl_metric_group}(.data, ...)

\method{select}{tbl_metric_group}(.data, ...)

\method{arrange}{tbl_metric_group}(.data, ...)

\method{mutate}{tbl_metric_group}(.data, ...)

\method{group_by}{tbl_metric_group}(.data, ...)

\method{summarise}{tbl_metric_group}(.data, ...)

\method{inner_join}{tbl_metric_group}(.data, ...)

\method{left_join}{tbl_metric_group}(.data, ...)

\method{right_join}{tbl_metric_group}(.data, ...)

\method{semi_join}{tbl_metric_group}(.data, ...)

\method{anti_join}{tbl_metric_group}(.data, ...)

\method{rename}{tbl_metric_group}(.data, ...)

\method{count}{tbl_metric_group}(.data, ...)

\method{transmute}{tbl_metric_group}(.data, ...)
}
\arguments{
\item{x}{For \code{as_tibble}, a \code{tbl_metric_group} object}

\item{...}{Arguments passed on to the appropriate dplyr verb}

\item{.data}{A \code{tbl_metric_group} object}
}
\description{
Metric dplyr S3 methods
}
