% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl-metric.R
\name{metric-methods}
\alias{metric-methods}
\alias{print.tbl_metric}
\title{S3 operators for metrics, including printing and coercing to a data frame}
\usage{
\method{print}{tbl_metric}(x, ...)
}
\arguments{
\item{x}{A tbl_metric}

\item{...}{Extra arguments, not used.}
}
\description{
S3 operators for metrics, including printing and coercing to a data frame
}
