% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl-metric.R
\name{metric-s3}
\alias{metric-s3}
\alias{as_tibble.tbl_metric}
\alias{filter.tbl_metric}
\alias{select.tbl_metric}
\alias{arrange.tbl_metric}
\alias{mutate.tbl_metric}
\alias{group_by.tbl_metric}
\alias{summarise.tbl_metric}
\alias{inner_join.tbl_metric}
\alias{left_join.tbl_metric}
\alias{right_join.tbl_metric}
\alias{semi_join.tbl_metric}
\alias{anti_join.tbl_metric}
\alias{rename.tbl_metric}
\alias{transmute.tbl_metric}
\alias{distinct.tbl_metric}
\title{Metric dplyr S3 methods}
\usage{
\method{as_tibble}{tbl_metric}(x)

\method{filter}{tbl_metric}(.data, ...)

\method{select}{tbl_metric}(.data, ...)

\method{arrange}{tbl_metric}(.data, ...)

\method{mutate}{tbl_metric}(.data, ...)

\method{group_by}{tbl_metric}(.data, ...)

\method{summarise}{tbl_metric}(.data, ...)

\method{inner_join}{tbl_metric}(.data, ...)

\method{left_join}{tbl_metric}(.data, ...)

\method{right_join}{tbl_metric}(.data, ...)

\method{semi_join}{tbl_metric}(.data, ...)

\method{anti_join}{tbl_metric}(.data, ...)

\method{rename}{tbl_metric}(.data, ...)

\method{transmute}{tbl_metric}(.data, ...)

\method{distinct}{tbl_metric}(.data, ...)
}
\arguments{
\item{x}{For as_data_frame, the}

\item{.data}{A \code{tbl_metric} object}

\item{...}{Arguments passed on to the appropriate dplyr metric}
}
\description{
Metric dplyr S3 methods
}
