% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl-metric-group.R
\name{create_metric_group}
\alias{create_metric_group}
\title{Create a metric group}
\usage{
create_metric_group(tbl, group_name = NULL, rmd_file = NULL)
}
\arguments{
\item{tbl}{A wide table of metrics}

\item{group_name}{An optional underscore separated string as group_name. If
not specified, the category and subcategory of the metrics in \code{tbl}
are concatenated to form the group_name. Specify a custom group_name only
if you are trying to save multiple metric groups in the same Rmd.}

\item{rmd_file}{The Rmd file that generated the compact metrics, which has
documentation for the metrics and dimensions stored in the YAML front matter.
If no Rmd file is given, it uses the currently running one.}
}
\description{
Create a metric group
}
