% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl-metric-group.R
\name{print.tbl_metric_group}
\alias{print.tbl_metric_group}
\title{Print method for tbl_metric_group}
\usage{
\method{print}{tbl_metric_group}(x, ...)
}
\arguments{
\item{x}{A tbl_metric_group object to \code{print}}

\item{...}{Additional parameters passed to \code{print}}
}
\description{
Print method for tbl_metric_group
}
