% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions.R
\name{remove_attribute_all}
\alias{remove_attribute_all}
\title{Remove attribute "All"}
\usage{
remove_attribute_all(tbl, ...)
}
\arguments{
\item{tbl}{A metric tbl in wide format, with one or more dimensions.}

\item{...}{Dimensions from which "All" should be removed, as bare names
or select helpers like \code{contains()}.}
}
\description{
This function removes the aggregate segment "All" for specified dimensions.
If no dimensions are specified, it removes the segment "All" from all
dimensions
}
\examples{
library(dplyr)

mtcars_by_cyl_gear <- mtcars \%>\%
  cross_by_dimensions(cyl, gear, vs) \%>\%
  summarize(avg_mpg = mean(mpg))

mtcars_by_cyl_gear \%>\%
  remove_attribute_all()

mtcars_by_cyl_gear \%>\%
  remove_attribute_all(cyl)

mtcars_by_cyl_gear \%>\%
  remove_attribute_all(-cyl, -gear)
}
