% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions.R
\name{remove_constant_dimensions}
\alias{remove_constant_dimensions}
\title{Remove dimensions with a constant level (single value)}
\usage{
remove_constant_dimensions(tbl, quietly = FALSE)
}
\arguments{
\item{tbl}{A metric tbl in wide format, with one or more dimensions.}

\item{quietly}{If FALSE (default), display a message about what columns are
being discarded.}
}
\description{
Use \code{remove_constant_dimensions} instead of \code{select} so the
removed dimension value is added to the metadata attribute.
}
\examples{
library(dplyr)
flights_nyc_avg_arr_delay \%>\%
  filter(origin == 'JFK') \%>\%
  remove_constant_dimensions()
}
