% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions.R
\name{var_names_dimensions}
\alias{var_names_dimensions}
\title{Get names of columns that are dimensions}
\usage{
var_names_dimensions(tbl)
}
\arguments{
\item{tbl}{A tbl_metric}
}
\description{
Any character of factor column not named date, value, period, or metric are
considered dimensions, as well as any columns ending in _id
}
