% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions.R
\name{var_names_not_dimensions}
\alias{var_names_not_dimensions}
\title{Get names of columns that are NOT dimensions}
\usage{
var_names_not_dimensions(tbl)
}
\arguments{
\item{tbl}{A tbl_metric}
}
\description{
Get names of columns that are NOT dimensions
}
