% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-metrics.R
\name{create_metrics}
\alias{create_metrics}
\title{Given a metric tbl and an Rmd file, turn into a named list of metric objects}
\usage{
create_metrics(..., rmd_file = NULL)
}
\arguments{
\item{...}{One or more metric tables in wide metric format: one column for each metric.}

\item{rmd_file}{The Rmd file that generated the compact metrics, which has
documentation for the metrics and dimensions stored in the YAML front matter.
If no Rmd file is given, it uses the currently running one.}
}
\value{
A named list of metric objects. Each of these has both the data and the metadata
(documentation, dimensions, owner, etc) to make an interactive visualization.
}
\description{
Given a metric tbl and an Rmd file, turn into a named list of metric objects
}
\examples{

# TODO

}
