% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{cross_join}
\alias{cross_join}
\title{Cross join two tables together, including all combinations of rows}
\usage{
cross_join(x, y, ...)
}
\arguments{
\item{x, y}{tbls to join}

\item{...}{additional arguments to be passed on to
\code{\link[dplyr]{full_join}}}
}
\description{
Locally, this is equivalent to tidyr::crossing.
}
\examples{
d1 <- dplyr::tibble(x = 1:3)
d2 <- dplyr::tibble(y = 1:2)
cross_join(d1, d2)
}
